# feat: improve syntax parser for pattern