--- a/pkg/logql/log/pattern/parser_test.go
+++ b/pkg/logql/log/pattern/parser_test.go
@@ -57,3 +57,14 @@ func Test_Parse(t *testing.T) {
 		require.Equal(t, tc.expected, actual)
 	}
 }
+
+var result expr
+
+func BenchmarkParseExpr(b *testing.B) {
+	var err error
+	b.ReportAllocs()
+	for i := 0; i < b.N; i++ {
+		result, err = parseExpr(`level=info <_> caller=main.go:107 msg="Starting Grafana Enterprise Traces" version="version=weekly-r138-f1920489, branch=weekly-r138, revision=f1920489"`)
+	}
+	require.NoError(b, err)
+}